pyproject.toml
[project]
name = "hybrid-agent"
version = "1.0.0"
description = "基于问题复杂度自动切换的多模型智能助手 + RAG 知识库"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.13",
    "dotenv>=0.9.9",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "sqlalchemy>=2.0.48",
    "sentence-transformers>=3.0.0",
    "sympy>=1.14.0",
    "streamlit>=1.55.0",
    "chromadb>=0.4.22",
    "langchain-chroma>=0.1.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "docx2txt>=0.8",
    "unstructured>=0.11.0",
    "markdown>=3.5.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-jose[cryptography]>=3.1.0",
    "passlib[bcrypt]>=1.7.4",
    "duckduckgo-search>=4.0.0",
    "ddgs>=9.0.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "rank-bm25>=0.2.2",
    "cachetools>=5.3.0",
    "structlog>=24.0.0",
    "prometheus-fastapi-instrumentator>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.12.0",
    "mypy>=1.17.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
hybrid-agent = "hybrid_agent.cli.main:run_cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["hybrid_agent*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
